backend/requirements.txt
openai
langchain
huggingface_hub
langchain-openai
cohere
google-cloud-storage
google-cloud-firestore
PyPDF2
numpy
pandas
matplotlib
sklearn
fastapi
uvicorn
